requirements.txt
streamlit
yfinance 
pandas 
plotly 
numpy 
talib 
websocket-client